requirements.txt
PyQt5==5.15.2
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.15.2.3.0.2
scikit-image==0.18.1
tensorflow==2.4.1
tensorflow-hub==0.11.0
numpy==1.19.5
matplotlib==3.3.4
Pillow==8.1.0
